Add option to exclude types with AutoMocker
