+++ bunit.AutoMocker/AutoMockerServiceProvider.cs
@@ -1,13 +1,18 @@
 ﻿using System;
+using System.Collections.Generic;
+using System.Linq;
 
 namespace Bunit.AutoMocker;
 
-public class AutoMockerServiceProvider(Moq.AutoMock.AutoMocker mocker) : IServiceProvider
+public class AutoMockerServiceProvider(Moq.AutoMock.AutoMocker mocker, params Type[] typesToExclude) : IServiceProvider
 {
-    private readonly Moq.AutoMock.AutoMocker Mocker = mocker;
+    private readonly HashSet<Type> _typesToExclude = typesToExclude.ToHashSet();
+    private readonly Moq.AutoMock.AutoMocker _mocker = mocker;
 
-    public object GetService(Type serviceType)
+    public object? GetService(Type serviceType)
     {
-        return Mocker.Get(serviceType);
+        return _typesToExclude.Contains(serviceType) 
+            ? null 
+            : _mocker.Get(serviceType);
     }
 }
