+++ bunit.AutoMocker/AutoMockerTestContext.cs
@@ -1,14 +1,16 @@
-﻿namespace Bunit.AutoMocker;
+﻿using System;
+
+namespace Bunit.AutoMocker;
 
 public abstract class AutoMockerTestContext : TestContext
 {
     protected readonly Moq.AutoMock.AutoMocker Mocker = new();
 
-    public AutoMockerTestContext() : this(new Moq.AutoMock.AutoMocker()) { }
+    public AutoMockerTestContext(params Type[] typesToExclude) : this(new Moq.AutoMock.AutoMocker(), typesToExclude) { }
 
-    public AutoMockerTestContext(Moq.AutoMock.AutoMocker mocker)
+    public AutoMockerTestContext(Moq.AutoMock.AutoMocker mocker, params Type[] typesToExclude)
     {
         Mocker = mocker;
-        Services.AddFallbackServiceProvider(new AutoMockerServiceProvider(Mocker));
+        Services.AddFallbackServiceProvider(new AutoMockerServiceProvider(Mocker, typesToExclude));
     }
 }
